requirements.txt
pandas==2.0.1
biopython
requests
bs4